Added search bar with semi working advanced search
The advanced search works for everything but the ingredients, I couldnt
figure out how to make it work for ingredients we may for the sake of
time and sanity have to drop that feature.  Also the views and
formatting all are terrible and need some editing its just super basic
layout.

Search bar right now is only on the recipe page.